refactor: move convert_camel_case_to_snake_case to utils
- Moved function from class_generator to ocp_resources/utils/utils.py
- This is a general utility function used by multiple modules
- Updated all imports to use the new location
- All tests still passing
- Avoids circular imports between resource.py and class_generator
